chore(comms): bump WsAccount wsdl to 1.06
bumps ws_account to latest (1.06) and changed service to use the
generated ws_account.ts
